implementation of seperation widget of dashboard_view
